GameController/requirements.txt
inputs==0.5
paho-mqtt==2.1.0
